+++ Bugsnag/Configuration/BugsnagConfiguration.m
@@ -88,7 +88,10 @@ static NSUserDefaults *userDefaults;
     [copy setEnabledBreadcrumbTypes:self.enabledBreadcrumbTypes];
     [copy setEnabledErrorTypes:self.enabledErrorTypes];
     [copy setEnabledReleaseStages:self.enabledReleaseStages];
+    copy.discardClasses = self.discardClasses;
     [copy setRedactedKeys:self.redactedKeys];
+    [copy setMaxPersistedEvents:self.maxPersistedEvents];
+    [copy setMaxPersistedSessions:self.maxPersistedSessions];
     [copy setMaxBreadcrumbs:self.maxBreadcrumbs];
     copy->_metadata = [[BugsnagMetadata alloc] initWithDictionary:[[self.metadata toDictionary] mutableCopy]];
     [copy setEndpoints:self.endpoints];
@@ -173,6 +176,8 @@ static NSUserDefaults *userDefaults;
     _enabledReleaseStages = nil;
     _redactedKeys = [NSSet setWithArray:@[@"password"]];
     _enabledBreadcrumbTypes = BSGEnabledBreadcrumbTypeAll;
+    _maxPersistedEvents = 12;
+    _maxPersistedSessions = 32;
     _maxBreadcrumbs = 25;
     _autoTrackSessions = YES;
     _sendThreads = BSGThreadSendPolicyAlways;
@@ -425,6 +430,46 @@ static NSUserDefaults *userDefaults;
 // MARK: - Properties: Getters and Setters
 // -----------------------------------------------------------------------------
 
+@synthesize maxPersistedEvents = _maxPersistedEvents;
+
+- (NSUInteger)maxPersistedEvents {
+    @synchronized (self) {
+        return _maxPersistedEvents;
+    }
+}
+
+- (void)setMaxPersistedEvents:(NSUInteger)maxPersistedEvents {
+    @synchronized (self) {
+        if (maxPersistedEvents >= 1 && maxPersistedEvents <= 100) {
+            _maxPersistedEvents = maxPersistedEvents;
+        } else {
+            bsg_log_err(@"Invalid configuration value detected. Option maxPersistedEvents "
+                        "should be an integer between 1-100. Supplied value is %lu",
+                        (unsigned long) maxPersistedEvents);
+        }
+    }
+}
+
+@synthesize maxPersistedSessions = _maxPersistedSessions;
+
+- (NSUInteger)maxPersistedSessions {
+    @synchronized (self) {
+        return _maxPersistedSessions;
+    }
+}
+
+- (void)setMaxPersistedSessions:(NSUInteger)maxPersistedSessions {
+    @synchronized (self) {
+        if (maxPersistedSessions >= 1 && maxPersistedSessions <= 100) {
+            _maxPersistedSessions = maxPersistedSessions;
+        } else {
+            bsg_log_err(@"Invalid configuration value detected. Option maxPersistedSessions "
+                        "should be an integer between 1-100. Supplied value is %lu",
+                        (unsigned long) maxPersistedSessions);
+        }
+    }
+}
+
 @synthesize maxBreadcrumbs = _maxBreadcrumbs;
 
 - (NSUInteger)maxBreadcrumbs {
@@ -443,6 +488,21 @@ static NSUserDefaults *userDefaults;
                         (unsigned long) maxBreadcrumbs);
         }
     }
+}
+
+- (BOOL)shouldDiscardErrorClass:(NSString *)errorClass {
+    for (id obj in self.discardClasses) {
+        if ([obj isKindOfClass:[NSString class]]) {
+            if ([obj isEqualToString:errorClass]) {
+                return YES;
+            }
+        } else if ([obj isKindOfClass:[NSRegularExpression class]]) {
+            if ([obj firstMatchInString:errorClass options:0 range:NSMakeRange(0, errorClass.length)]) {
+                return YES;
+            }
+        }
+    }
+    return NO;
 }
 
 /**